# specs/v1/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "nanocld",
    "description": "",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": ""
    },
    "version": "0.1.3"
  },
  "paths": {
    "/cargoes": {
      "get": {
        "tags": [
          "cargo"
        ],
        "summary": "List cargo",
        "description": "List cargo\n",
        "operationId": "list_cargo",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo are stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cargo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CargoItem"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "cargo"
        ],
        "summary": "Create new cargo",
        "description": "Create new cargo\n",
        "operationId": "create_cargo",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo will be stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CargoPartial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "New cargo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CargoItem"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
    "/clusters": {
      "get": {
        "tags": [
          "cluster"
        ],
        "summary": "List all cluster",
        "description": "List all cluster\n",
        "operationId": "list_cluster",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Namespace to add cluster in if empty we use 'default' as value",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of cluster for given namespace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterItem"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error"
          },
          "404": {
            "description": "Namespace name not valid"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "cluster"
        ],
        "summary": "Create new cluster",
        "description": "Create new cluster\n",
        "operationId": "create_cluster",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Namespace to add cluster in if empty we use 'default' as value",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterPartial"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Fresh created cluster",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterItem"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error"
          },
          "404": {
            "description": "Namespace name not valid"
          }
        },
        "deprecated": false
      }
    },
    "/clusters/count": {
      "get": {
        "tags": [
          "cluster"
        ],
        "summary": "Count cluster",
        "description": "Count cluster\n",
        "operationId": "count_cluster",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "description": "Name of the namespace where the cargo is stored",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generic delete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgGenericCount"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
    "/clusters/{name}": {
      "delete": {
        "tags": [
          "cluster"
        ],
        "summary": "Delete cluster by it's name",
        "description": "Delete cluster by it's name\n",
        "operationId": "delete_cluster_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the cluster",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Namespace to add cluster in if empty we use 'default' as value",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Fresh created cluster",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterItem"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/clusters/{name}/inspect": {
      "get": {
        "tags": [
          "cluster"
        ],
        "summary": "Inspect cluster by it's name",
        "description": "Inspect cluster by it's name\n",
        "operationId": "inspect_cluster_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the cluster",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Namespace to add cluster in if empty we use 'default' as value",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cluster information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterItemWithRelation"
                }
              }
            }
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "id name or namespace name not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/clusters/{name}/join": {
      "post": {
        "tags": [
          "cluster"
        ],
        "summary": "join cargo inside a cluster",
        "description": "join cargo inside a cluster\n",
        "operationId": "join_cargo_to_cluster",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the cluster",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Namespace to add cluster in if empty we use 'global' as value",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClusterJoinBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cargo joinned successfully"
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Cluster name of namespace invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
    "/clusters/{name}/start": {
      "post": {
        "tags": [
          "cluster"
        ],
        "summary": "Start all cargo inside cluster",
        "description": "Start all cargo inside cluster\n",
        "operationId": "start_cluster_by_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Name of the cluster",
            "required": true,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "in": "query",
            "description": "Namespace to add cluster in if empty we use 'global' as value",
            "required": false,
            "deprecated": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cargos have been started"
          },
          "400": {
            "description": "Generic database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Cluster name of namespace invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "Api Error Structure that server send to client",
        "required": [
          "msg"
        ],
        "properties": {
          "msg": {
            "type": "string"
          }
        }
      },
      "CargoItem": {
        "type": "object",
        "description": "Cargo item is an definition to container create image and start them",
        "required": [
          "key",
          "namespace_name",
          "name",
          "image_name",
          "binds",
          "replicas"
        ],
        "properties": {
          "binds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dns_entry": {
            "type": "string"
          },
          "domainname": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "image_name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "namespace_name": {
            "type": "string"
          },
          "replicas": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CargoPartial": {
        "type": "object",
        "description": "Cargo partial",
        "required": [
          "name",
          "image_name"
        ],
        "properties": {
          "binds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dns_entry": {
            "type": "string"
          },
          "domainname": {
            "type": "string"
          },
          "environnements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hostname": {
            "type": "string"
          },
          "image_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "replicas": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ClusterItem": {
        "type": "object",
        "description": "Cluster used to encapsulate networks",
        "required": [
          "key",
          "name",
          "namespace",
          "proxy_templates"
        ],
        "properties": {
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "namespace": {
            "type": "string"
          },
          "proxy_templates": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ClusterJoinBody": {
        "type": "object",
        "required": [
          "cargo",
          "network"
        ],
        "properties": {
          "cargo": {
            "type": "string"
          },
          "network": {
            "type": "string"
          }
        }
      },
...
      "ClusterPartial": {
        "type": "object",
        "description": "Partial cluster",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "proxy_templates": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}